reservation pas deux fois daffilé

+++ htdocs/bimpcore/objects/BimpFile.class.php
@@ -1,7 +1,10 @@
 <?php
+        require_once DOL_DOCUMENT_ROOT.'/core/lib/files.lib.php';
 
 class BimpFile extends BimpObject
 {
+    
+    private $dontRemove = false;
 
     public static $types = array(
         'pdf' => array('label' => 'PDF', 'icon' => 'file-pdf-o'),
@@ -44,6 +47,8 @@ class BimpFile extends BimpObject
     // Traitement des fichiers: 
     public function uploadFile()
     {
+        
+        //$this->getFileUrl();die;
         $errors = array();
 
         $file = $_FILES['file'];
@@ -58,20 +63,15 @@ class BimpFile extends BimpObject
         $_FILES['file']['name'] = $newFileName;
         
         
-        $modulepart2 = $this->getData('files_dir');
-        
-        
-        $tabT = json_decode(GETPOST("associations_params"));
-        $id = $tabT[0]->id_object;
-        
-
-        $upload_dir = DOL_DATA_ROOT."/bimpcore/".$modulepart2."/".$id;
-        
-        require_once DOL_DOCUMENT_ROOT.'/core/lib/files.lib.php';
-        
-        $ret = dol_add_file_process($upload_dir, 0, 0, 'file');
-        if(!$ret)
-                $errors[] = "Fichier non enregistré";
+        if(is_file($this->getFilePath().$newFileName)){
+                $errors[] = "Le Fichier existe déja";
+                $this->dontRemove = true;
+        }
+        else{
+            $ret = dol_add_file_process($this->getFilePath(), 0, 0, 'file');
+            if(!$ret)
+                    $errors[] = "Fichier non enregistré";
+        }
 
         // Ajouter toute erreur à $errors (texte) 
         // Si pour une raison ou autre la taille du fichier est modifié, faire juste: $this->set('file_size', $new_size)
@@ -79,23 +79,35 @@ class BimpFile extends BimpObject
 
         return $errors;
     }
-
-    public function getFilePath()
-    {
-        
+    
+    private function getFilePath(){
+        $modulepart2 = $this->getData('files_dir');
+        return DOL_DATA_ROOT."/bimpcore/".$modulepart2."/".$this->getId()."/";
+    }
+    
+    private function getId(){
+        $tabT = json_decode(GETPOST("associations_params"));
+        $id = $tabT[0]->id_object;
+        return $id;
     }
 
     public function getFileUrl()
     {
-        echo 'jjjjj';
+        $file = $_FILES['file']["name"];
+        
+        $modulepart2 = $this->getData('files_dir');
+        echo DOL_URL_ROOT.'/document.php?modulepart=bimpcore&file='.$modulepart2."/".$this->getId()."/".$file;
     }
 
     public function removeFile()
     {
         $errors = array();
         
+        $file = $_FILES['file']["name"];
         
-        dol_remove_file_process($filenb);
+        
+        
+        dol_delete_file($this->getFilePath()."/".$file);
         
         return $errors;
     }
@@ -129,7 +141,7 @@ class BimpFile extends BimpObject
             $errors = parent::create();
         }
 
-        if (count($errors)) {
+        if (count($errors)  && !$this->dontRemove) {
             $this->removeFile();
         }
